Extract common logic in SetupCommandAsync
