+++ Moq.Dapper/DbConnectionMockExtensions.cs
@@ -45,43 +45,11 @@ namespace Moq.Dapper
         }
 
         private static ISetup<DbConnection, Task<TResult>> SetupQueryAsync<TResult>(Mock<DbConnection> mock) =>
-            SetupCommandAsync<TResult>(mock, (commandMock, result) =>
+            DbCommandSetup.SetupCommandAsync<TResult, DbConnection>(mock, (commandMock, result) =>
             {
                 commandMock.Protected()
                            .Setup<Task<DbDataReader>>("ExecuteDbDataReaderAsync", ItExpr.IsAny<CommandBehavior>(), ItExpr.IsAny<CancellationToken>())
                            .ReturnsAsync(() => DbDataReaderFactory.DbDataReader(result));
             });
-
-        private static ISetup<DbConnection, Task<TResult>> SetupCommandAsync<TResult>(Mock<DbConnection> mock, Action<Mock<DbCommand>, Func<TResult>> mockResult)
-        {
-            var setupMock = new Mock<ISetup<DbConnection, Task<TResult>>>();
-
-            var result = default(TResult);
-            
-            setupMock.Setup(setup => setup.Returns(It.IsAny<Func<Task<TResult>>>()))
-                     .Callback<Func<Task<TResult>>>(r => result = r().Result);
-
-            var commandMock = new Mock<DbCommand>();
-
-            commandMock.Protected()
-                       .SetupGet<DbParameterCollection>("DbParameterCollection")
-                       .Returns(new Mock<DbParameterCollection>().Object);
-
-            commandMock.Protected()
-                       .Setup<DbParameter>("CreateDbParameter")
-                       .Returns(new Mock<DbParameter>().Object);
-
-            mockResult(commandMock, () => result);
-
-            mock.As<IDbConnection>()
-                .Setup(m => m.CreateCommand())
-                .Returns(commandMock.Object);
-
-            mock.Protected()
-                .Setup<DbCommand>("CreateDbCommand")
-                .Returns(commandMock.Object);
-
-            return setupMock.Object;
-        }
     }
 }
